gRPC stream ReadWriteCloser implementation
Wraps a grpc stream (ClientStream/ServerStream) around
a ReadWriteCloser implementation.
Users must provide functions that now how to Read/Write bytes
from/to the underlying stream message type and a function to create
a message that the other side understands as a close.
The StreamReadWriteCloser type uses this to provide the Read/Write/Close
methods.

Using this makes it easy to overlay io operation on top of the stream like
stdin/stdout redirection, port forwarding ...

Change-Id: I95d93ff502742bff464f57c7c5e8b423206bca5b