Fix broken builds
- Add flags in .bazelrc so that broken changes introduced in recent
version of bazel don't break dependencies that still rely on old
behavior.

- Fetch io_grpc_grpc_java before everything else. Both go rules and
java_grpc depend on protobuf repo, however the go rules version depends
on a version which is broken for java
(https://github.com/bazelbuild/rules_go/pull/2105 addresses this).
Until that change is not submitted we need to force the java dep to be
downloaded first. Note that we cant revert to an older go rules version
since they are not compatible with newer bazel binaries (and no way to
work around via bazelrc).

Correct soulution is to have a repositories.bzl that loads common
dependencies up front in order to avoid this issues.