Allow running server forwarder over multiplex connections.
Introduced forkfd a new binary that listens for incoming connections to be
multiplexed and starts the server forking the process with a dup of the
connection fd. The forwarder can then connect to the server over
multiplexed streams on that endpoint.

forkfd opens a socket that can only handle a single stream at a time.
Once a connection is established the waterfall process is forked with
access to that fd. Waterall uses that connection to start a
Multiplexer service which is then used to start the waterfall server.
forfd blocks until the waterfall process is done since the communication
channel opened by forkfs is only suppose to ever have a single active
connection.